+++ src/src_c/Pimpact_ConvDiff2.cxx
@@ -77,9 +77,9 @@ int main( int argi, char** argv ) {
       if( space()->rankST()==0 ) {
         std::string fname = "raki.txt";
         if( 3==dirx )
-          fname.insert( 4, std::to_string( (int)8 ) );
+          fname.insert( 4, std::to_string( (long long)8 ) );
         else
-          fname.insert( 4, std::to_string( dirx+diry*2+3 ) );
+          fname.insert( 4, std::to_string( (long long)dirx+diry*2+3 ) );
         phifile.open( fname, std::ofstream::out);
       }
