+++ src/src_c/Pimpact_ConvDiffMG.cxx
@@ -114,9 +114,9 @@ int main( int argi, char** argv ) {
       if( space()->rankST()==0 ) {
         std::string fname = "phin.txt";
         if( 3==dirx )
-          fname.insert( 4, std::to_string( (int)8 ) );
+          fname.insert( 4, std::to_string( (long long)8 ) );
         else
-          fname.insert( 4, std::to_string( dirx+diry*2+3 ) );
+          fname.insert( 4, std::to_string( (long long)dirx+diry*2+3 ) );
         phifile.open( fname, std::ofstream::out);
       }
 
@@ -160,7 +160,7 @@ int main( int argi, char** argv ) {
 
         std::ofstream ofs;
         std::string filename = "GS.txt";
-        filename.insert( 2, std::to_string(phii) );
+        filename.insert( 2, std::to_string( (long long)phii) );
 
         if( space()->rankST()==0 )
           ofs.open(filename, std::ofstream::out);
